+++ index.js
@@ -1,6 +1,5 @@
 // export { default as updateMeta } from './lib/meta/patchMeta.js'
 
-export { default as getState } from './lib/redux/client/getState.js'
 export { default as getHttpClient } from './lib/redux/client/getHttpClient.js'
 
 export {
@@ -38,7 +37,10 @@ export {
 
 export {
 	Redirect,
-	useRouter,
+	useRouter
+} from './lib/router/index.js'
+
+export {
 	goto,
 	redirect,
 	pushLocation,
@@ -46,12 +48,18 @@ export {
 	goBack,
 	goBackTwoPages,
 	goForward
-} from './lib/router/index.js'
+} from './lib/router/actions.js'
 
 export { default as useNavigationStartEffect } from './lib/router/useNavigationStartEffect.js'
 export { default as useNavigationEndEffect } from './lib/router/useNavigationEndEffect.js'
+export { default as useNavigationLocation } from './lib/redux/navigation/useNavigationLocation.js'
 export { default as useSelectorForLocation } from './lib/redux/useSelectorForLocation.js'
 export { default as useLocation } from './lib/router/useLocation.js'
+export { default as useLocationHistory } from './lib/router/useLocationHistory.js'
+export { default as useGoBack } from './lib/router/useGoBack.js'
+export { default as useGoForward } from './lib/router/useGoForward.js'
+export { default as useNavigate } from './lib/router/useNavigate.js'
+export { default as useRedirect } from './lib/router/useRedirect.js'
 export { default as useLoading } from './lib/router/useLoading.js'
 export { default as useRoute } from './lib/router/useRoute.js'